Replaced BigInteger with Int

+++ lib/src/main/java/com/mastercard/gateway/ats/domain/ATSErrorType.kt
@@ -2,7 +2,6 @@ package com.mastercard.gateway.ats.domain
 
 import org.simpleframework.xml.Attribute
 import org.simpleframework.xml.Root
-import java.math.BigInteger
 
 @Suppress("PLATFORM_CLASS_MAPPED_TO_KOTLIN")
 @Root(name = "ATSErrorType", strict = false)
@@ -15,6 +14,6 @@ class ATSErrorType {
     @field:Attribute(name = "Description", required = false)
     var description: String? = null
     @field:Attribute(name = "Severity", required = false)
-    var severity: BigInteger? = null
+    var severity: Int? = null
 
 }
