+++ lib/src/main/java/com/mastercard/gateway/ats/domain/DeviceResponse.kt
@@ -2,35 +2,32 @@ package com.mastercard.gateway.ats.domain
 
 import org.simpleframework.xml.Attribute
 import org.simpleframework.xml.Element
-import org.simpleframework.xml.Order
 import org.simpleframework.xml.Root
-
 import java.math.BigDecimal
-import java.math.BigInteger
 
 @Root(name = "DeviceResponse", strict = false)
 class DeviceResponse : ATSMessage {
 
-    @field:Element(name = "Output", required=false)
+    @field:Element(name = "Output", required = false)
     var output: List<Output>? = null
-    @field:Element(name = "Input", required=false)
+    @field:Element(name = "Input", required = false)
     var input: DeviceResponse.Input? = null
-    @field:Element(name = "EventResult", required=false)
+    @field:Element(name = "EventResult", required = false)
     var eventResult: DeviceResponse.EventResult? = null
-    @field:Element(name = "ErrorDetail", required=false)
+    @field:Element(name = "ErrorDetail", required = false)
     var errorDetail: ATSErrorDetailType? = null
 
-    @field:Attribute(name = "ApplicationSender", required=false)
+    @field:Attribute(name = "ApplicationSender", required = false)
     var applicationSender: String? = null
-    @field:Attribute(name = "WorkstationID", required=false)
+    @field:Attribute(name = "WorkstationID", required = false)
     var workstationID: String? = null
-    @field:Attribute(name = "TerminalID", required=false)
+    @field:Attribute(name = "TerminalID", required = false)
     var terminalID: String? = null
-    @field:Attribute(name = "POPID", required=false)
+    @field:Attribute(name = "POPID", required = false)
     var popid: String? = null
-    @field:Attribute(name = "SequenceID", required=false)
+    @field:Attribute(name = "SequenceID", required = false)
     var sequenceID: Int? = null
-    @field:Attribute(name = "ReferenceRequestID", required=false)
+    @field:Attribute(name = "ReferenceRequestID", required = false)
     var referenceRequestID: String? = null
     @field:Attribute(name = "RequestType", required = true)
     lateinit var requestType: DeviceRequestType
@@ -42,88 +39,88 @@ class DeviceResponse : ATSMessage {
 
     class EventResult {
 
-        @field:Element(name = "Dispenser", type = Short::class, required=false)
+        @field:Element(name = "Dispenser", type = Short::class, required = false)
         var dispenser: List<Short>? = null
-        @field:Element(name = "ProductCode", required=false)
-        var productCode: List<BigInteger>? = null
-        @field:Element(name = "ModifiedRequest", required=false)
+        @field:Element(name = "ProductCode", required = false)
+        var productCode: List<Int>? = null
+        @field:Element(name = "ModifiedRequest", required = false)
         var modifiedRequest: CardRequestType? = null
-        @field:Element(name = "TotalAmount", required=false)
+        @field:Element(name = "TotalAmount", required = false)
         var totalAmount: TotalAmountType? = null
-        @field:Element(name = "SaleItem", required=false)
+        @field:Element(name = "SaleItem", required = false)
         var saleItem: List<SaleItemType>? = null
-        @field:Element(name = "Acquirer", required=false)
+        @field:Element(name = "Acquirer", required = false)
         var acquirer: Acquirer? = null
-        @field:Element(name = "AuthResponse", required=false)
+        @field:Element(name = "AuthResponse", required = false)
         var authResponse: AuthResponse? = null
-        @field:Element(name = "TransactionLimits", required=false)
+        @field:Element(name = "TransactionLimits", required = false)
         var transactionLimits: TransactionLimits? = null
-        @field:Element(name = "OACentreDetails", required=false)
+        @field:Element(name = "OACentreDetails", required = false)
         var oaCentreDetails: OACentreDetails? = null
 
         class Acquirer {
-            @field:Attribute(name = "MerchantID", required=false)
+            @field:Attribute(name = "MerchantID", required = false)
             var merchantID: String? = null
-            @field:Attribute(name = "AcquirerID", required=false)
+            @field:Attribute(name = "AcquirerID", required = false)
             var acquirerID: String? = null
-            @field:Attribute(name = "MerchantReference", required=false)
+            @field:Attribute(name = "MerchantReference", required = false)
             var merchantReference: String? = null
-            @field:Attribute(name = "CreditPlan", required=false)
+            @field:Attribute(name = "CreditPlan", required = false)
             var creditPlan: String? = null
         }
 
         class AuthResponse {
-            @field:Element(name = "AuthARC", required=false)
+            @field:Element(name = "AuthARC", required = false)
             var authARC: String? = null
-            @field:Element(name = "AuthCode", required=false)
+            @field:Element(name = "AuthCode", required = false)
             var authCode: String? = null
-            @field:Element(name = "AuthResult", required=false)
+            @field:Element(name = "AuthResult", required = false)
             var authResult: String? = null
-            @field:Element(name = "ICC", required=false)
+            @field:Element(name = "ICC", required = false)
             var icc: ICCType? = null
         }
 
         class OACentreDetails {
-            @field:Element(name = "OACentreName", required=false)
+            @field:Element(name = "OACentreName", required = false)
             var oaCentreName: String? = null
-            @field:Element(name = "RefundOnline", required=false)
+            @field:Element(name = "RefundOnline", required = false)
             var refundOnline: OARefundOnline? = null
-            @field:Element(name = "RefundOnlineCtls", required=false)
+            @field:Element(name = "RefundOnlineCtls", required = false)
             var refundOnlineCtls: OARefundOnline? = null
-            @field:Element(name = "ReversalMode", required=false)
+            @field:Element(name = "ReversalMode", required = false)
             var reversalMode: OAReversalMode? = null
-            @field:Element(name = "ReverseOnline", required=false)
+            @field:Element(name = "ReverseOnline", required = false)
             var reverseOnline: OAReverseOnline? = null
-            @field:Element(name = "ICCNoTrack2", required=false)
+            @field:Element(name = "ICCNoTrack2", required = false)
             var iccNoTrack2: OAICCNoTrack2? = null
         }
 
         class TransactionLimits {
-            @field:Element(name = "AuthFlags", required=false)
+            @field:Element(name = "AuthFlags", required = false)
             var authFlags: String? = null
-            @field:Element(name = "LimitFloor", required=false)
+            @field:Element(name = "LimitFloor", required = false)
             var limitFloor: BigDecimal? = null
-            @field:Element(name = "LimitPostComms", required=false)
+            @field:Element(name = "LimitPostComms", required = false)
             var limitPostComms: BigDecimal? = null
-            @field:Element(name = "Ceiling", required=false)
+            @field:Element(name = "Ceiling", required = false)
             var ceiling: BigDecimal? = null
-            @field:Element(name = "CeilingCDCVM", required=false)
+            @field:Element(name = "CeilingCDCVM", required = false)
             var ceilingCDCVM: BigDecimal? = null
-            @field:Element(name = "CeilingCash", required=false)
+            @field:Element(name = "CeilingCash", required = false)
             var ceilingCash: BigDecimal? = null
-            @field:Element(name = "CeilingFPA", required=false)
+            @field:Element(name = "CeilingFPA", required = false)
             var ceilingFPA: BigDecimal? = null
-            @field:Element(name = "ReadTLV", required=false)
-            var readTLV: BigInteger? = null
+            @field:Element(name = "ReadTLV", required = false)
+            var readTLV: Int? = null
         }
 
     }
 
     class Input {
 
-        @field:Element(name = "SecureData", required=false)
+        @field:Element(name = "SecureData", required = false)
         var secureData: List<SecureDataFlow>? = null
-        @field:Element(name = "InputValue", required=false)
+        @field:Element(name = "InputValue", required = false)
         var inputValue: InputValueType? = null
         @field:Attribute(name = "InDeviceTarget", required = true)
         lateinit var inDeviceTarget: DeviceType
